Seperate ThingsBoard into seperate API implementations
