+++ keywords.txt
@@ -110,7 +110,6 @@ Measure_Json    KEYWORD2
 #######################################
 
 THINGSBOARD_ENABLE_STL	LITERAL1
-THINGSBOARD_ENABLE_OTA  LITERAL1
 THINGSBOARD_ENABLE_DYNAMIC  LITERAL1
 THINGSBOARD_ENABLE_DEBUG    LITERAL1
 THINGSBOARD_ENABLE_STREAM_UTILS LITERAL1
