+++ src/SDCard_Updater.cpp
@@ -1,8 +1,6 @@
 // Header include.
 #include "SDCard_Updater.h"
 
-#if THINGSBOARD_ENABLE_OTA
-
 // Library include.
 #include <stdio.h>
 
@@ -39,5 +37,3 @@ void SDCard_Updater::reset() {
 bool SDCard_Updater::end() {
     return remove(m_path) == 0;
 }
-
-#endif // THINGSBOARD_ENABLE_OTA
